OCP-996 - Eric's updates from Azure repo
This includes UIs for forgot password/username and password reset by verification code. Still needs some TLC, we will get to that later, for now, it at least will fire off the requests - but the message sender would need to be configured for blob storage for us to pull those verification codes out.